+++ src/jaeger/actor/commands/fvc.py
@@ -148,6 +148,11 @@ async def expose(
     help="Comma-separated ZB orders to use for the FVC transformation.",
 )
 @click.option(
+    "--rotator-reference-position",
+    type=float,
+    help="Rotator reference position for fiducial fitting.",
+)
+@click.option(
     "--no-write-summary",
     is_flag=True,
     help="Does not try to write a confSummaryF file.",
@@ -362,6 +367,7 @@ async def take_fvc_loop(
     no_write_summary: bool = False,
     configuration: BaseConfiguration | None = None,
     polids: list[int] | str | None = None,
+    rot_ref_angle: float | None = None,
 ):
     """Helper to take an FVC loop that can be called externally."""
 
@@ -445,6 +451,7 @@ async def take_fvc_loop(
                 polids=polids,
                 centroid_method=centroid_method,
                 use_new_invkin=use_invkin,
+                rot_ref_angle=rot_ref_angle,
                 # loop=asyncio.get_running_loop(),  # Disable for now
             )